update map activity design

+++ app/src/main/java/com/empathy/empathy_android/extensions/SupportExtensions.kt
@@ -3,8 +3,8 @@ package com.empathy.empathy_android.extensions
 import android.content.res.Resources
 
 
-val Int.toPx: Int
-    get() = (this * Resources.getSystem().displayMetrics.density).toInt()
+val Int.toPx: Float
+    get() = (this * Resources.getSystem().displayMetrics.density).toFloat()
 
-val Int.toDp: Int
-    get() = (this / Resources.getSystem().displayMetrics.density).toInt()
+val Int.toDp: Float
+    get() = (this / Resources.getSystem().displayMetrics.density).toFloat()
